+++ src/config/index.js
@@ -1,1 +1,3 @@
 export * from './config'
+
+export const BRANDING_KEY = '_branding'