+++ src/usecases/branding/fetch.js
@@ -1,0 +1,3 @@
+import { fetchBranding } from '@/lib/intercom/fetchBranding'
+
+export default async () => fetchBranding()